% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{data}
\name{APIinfo}
\alias{APIinfo}
\title{tsentiment environment}
\format{
An object of class \code{environment} of length 4.
}
\usage{
APIinfo
}
\arguments{
\item{APIinfo}{Environment}
}
\description{
tsentiment environment
}
\keyword{datasets}
