% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileConfirmation.R
\name{checkConfirmForUSer}
\alias{checkConfirmForUSer}
\title{User file permission checker}
\usage{
checkConfirmForUSer()
}
\value{
String
}
\description{
User file permission checker
}
\examples{
\dontrun{
checkConfirmForUser()
}
}
