% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPageLimit.R
\name{checkPageLimit}
\alias{checkPageLimit}
\title{Check page limit}
\usage{
checkPageLimit(q)
}
\arguments{
\item{q}{is quantity for tweet}
}
\value{
integer
}
\description{
This function looping for pagination calling while looping
}
\examples{
q <- 300
checkPageLimit(q)
}
