% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRversion.R
\name{checkVersionForSentiment}
\alias{checkVersionForSentiment}
\title{Check r Version}
\usage{
checkVersionForSentiment()
}
\value{
boolean
}
\description{
This function checking using current R version
}
\examples{
checkVersionForSentiment()
}
