% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterDataCleaner.R
\name{cleanFetchedTweet}
\alias{cleanFetchedTweet}
\title{Clean raw data}
\usage{
cleanFetchedTweet(fetchedTweet = NULL)
}
\arguments{
\item{fetchedTweet}{Fetched tweet has a lot of different characters}
}
\value{
table
}
\description{
This function clean fetched tweets
}
