% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{getBarSentiment}
\alias{getBarSentiment}
\title{Export bar plot}
\usage{
getBarSentiment(tweet = NULL)
}
\arguments{
\item{tweet}{Cleaned tweet data}
}
\value{
file
}
\description{
This function export a bar plot with analysed data
}
