% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisonCloud.R
\name{getCloudSentiment}
\alias{getCloudSentiment}
\title{Export word cloud}
\usage{
getCloudSentiment(text)
}
\arguments{
\item{text}{Cleaned tweet data}
}
\value{
file
}
\description{
This function export a word cloud with analysed data
}
