% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTweet.R
\name{getTweet}
\alias{getTweet}
\title{Get Tweet}
\usage{
getTweet(fetchParams)
}
\arguments{
\item{fetchParams}{is variable}
}
\value{
JSON
}
\description{
This function only get tweet from tsentiment api
}
\examples{
\dontrun{
fetchParams <- list(headers = headers,params = params,url = APIinfo$url)
getTweet(fetchParams)
}
}
