% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextTweetLooper.R
\name{nextTweetFetcher}
\alias{nextTweetFetcher}
\title{Looping after first tweet fetching}
\usage{
nextTweetFetcher(nextInfo = NULL, nextParams = NULL, q = NULL)
}
\arguments{
\item{nextInfo}{sending from tweetFetcher function with final form}

\item{nextParams}{required after the first tweet fetch process}

\item{q}{is quantity, so its looping pagination limit}
}
\value{
String
}
\description{
This function should be used at the start of each session
}
\examples{
headers <- c(`Authorization` = sprintf('Bearer \%s', "YOUR BEARER TOKEN"))
params <- list(`query` = paste("Your-query", "lang:en"),`max_results` = 100,
`tweet.fields` = 'created_at,lang,conversation_id',`page` = 1)
nextInfo <- list("headers" = headers,"params" = params, url = "url")
nextParams <- list(`query` = nextInfo$params$query,
`next_token` = "After fetching take a token for next tweet fetch")
\dontrun{
nextTweetFetcher(nextInfo,nextParams,q = nextInfo$params$page)
}
}
