% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetFetcher.R
\name{tweetFetcher}
\alias{tweetFetcher}
\title{Fetch Tweets}
\usage{
tweetFetcher(APIinfo)
}
\arguments{
\item{APIinfo}{Entered API information}
}
\value{
list
}
\description{
This function prepare http information for fetching process and working together sub functin
}
\examples{
APIinfo <- list(BEARER_TOKEN = "DSEFS55SSS",query = "binance",page = 300)
tweetFetcher(APIinfo)
}
