% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSV.R
\name{writeToCSV}
\alias{writeToCSV}
\title{Create CSV}
\usage{
writeToCSV(tweetData)
}
\arguments{
\item{tweetData}{fetched first dirty tweet data}
}
\value{
file
}
\description{
This function create CSV file with parameters
}
\examples{
\dontrun{
writeToCSV()
}
}
