% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalysis.R
\name{getAnalysis}
\alias{getAnalysis}
\title{Start analysis after defined Twitter API information}
\usage{
getAnalysis()
}
\value{
file
}
\description{
This function prepare API information and start analysis
}
\examples{
\dontrun{
clearPrevious()
getCloudSentiment()
getBarSentiment()
}
}
