% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetFetcher.R
\name{tweetFetcher}
\alias{tweetFetcher}
\title{Fetch Tweets}
\usage{
APIinfo
}
\arguments{
\item{APIinfo}{is environment}
}
\value{
list
}
\description{
This function prepare http information for fetching process and working together sub functin
}
\examples{
APIinfo <- list(BEARER_TOKEN = "token",query = "love",page = 100)
tweetFetcher()
}
